backend/requirements.txt
flask==2.2
werkzeug==2.2
sqlalchemy==2.0.0
flask-sqlalchemy==3.0.0
psycopg2-binary==2.9.5
python-dotenv==0.19.2
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
